+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to `laravel-optimus` will be documented in this file.
 
+## [2.0.0] - 2017-09-09
+
+### Added
+
+- Laravel 5.5 support
+- Service provider and facade auto-discovery
+
+### Changed
+
+- Namespace `Cog\Optimus` renamed to `Cog\Laravel\Optimus`
+
 ## 1.0.0 - 2017-01-01
 
 Initial release
+
+[2.0.0]: https://github.com/cybercog/laravel-optimus/compare/1.0.0...2.0.0